+++ entity/src/main/java/org/grnet/cat/entities/MotivationAssessment.java
@@ -1,7 +1,14 @@
 package org.grnet.cat.entities;
 
 
-import jakarta.persistence.*;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,13 +19,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * This entity represents the Assessment table in database.
+ * This entity represents the Motivation Assessment table in database.
  *
  */
 @Entity
 @Getter
 @Setter
-public class Assessment {
+public class MotivationAssessment {
 
     /**
      * As id
@@ -29,10 +36,6 @@ public class Assessment {
 
     @Column(name = "assessment_doc", columnDefinition = "json")
     private String assessmentDoc;
-
-    @ManyToOne
-    @JoinColumn(name = "template_id", referencedColumnName = "id")
-    private Template template;
 
     @ManyToOne
     @JoinColumn(name = "validation_id", referencedColumnName = "id")
@@ -52,7 +55,7 @@ public class Assessment {
     @Column(name = "updated_by")
     private String updatedBy;
 
-    @OneToMany(mappedBy = "assessment",
+    @OneToMany(mappedBy = "motivationAssessment",
                 cascade = CascadeType.ALL,
                 orphanRemoval = true)
     private List<Comment> comments = new ArrayList<>();